Docstore: fix UX when submitting final page of incomplete docstore
Check the instance for any incomplete pages when submitting the final
page rather than assuming prior pages are complete. If incomplete pages
exist, render an 'incomplete' page, with an overview of the pages and
their status, to make things clearer for users/don't try to go to a page
we shouldn't yet.