Export filtering: fix filtering when fact name contains capital letters
When fact names contain capital letters, export filtering previously
didn’t work - the name was getting converted to lowercase for saving as
the name on the filter. This lowercase version was then used to check
the value of the named fact for each row. When the fact name contained
an uppercase character, the names wouldn’t match, so it would not be
found, and therefore wouldn’t be filtered on